requirements.txt
psycopg2-binary
requests
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv

fastapi
uvicorn
